requirements.txt
mss
keyboard
opencv-python
numpy
torch
torchvision
torchaudio
pillow
matplotlib
scikit-learn
seaborn
tqdm
